+++ src/components/spartan/SSectionTitle/SSectionTitle.vue
@@ -1,12 +1,9 @@
 <script setup lang="ts">
-withDefaults(
-    defineProps<{
-        as: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
-    }>(),
-    {
-        as: 'h3',
-    },
-);
+import type { TSectionTitleProps } from './types';
+
+withDefaults(defineProps<TSectionTitleProps>(), {
+    as: 'h3',
+});
 </script>
 
 <template>