+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- `SSectionTitle` now exports his types.
 - `SSectionDescription` now exports his types.
 - `SRadioGroupItem` now exports his types.
 - `SRadioGroup` now exports his types.